Move language selector into individual component.
